Refactor watering controller: implement water alarm functionality and update imports across the application
