+++ lib/widgets/water_widget/change_water_modal/change_time_modal.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:hexcolor/hexcolor.dart';
-import 'package:iot_plant_control/controller/water_controller.dart';
+import 'package:iot_plant_control/controller/watering_controller/water_controller.dart';
 import 'package:iot_plant_control/models/water_time.dart';
 import 'package:iot_plant_control/widgets/water_widget/change_water_modal/button_confirm.dart';
 import 'package:iot_plant_control/widgets/water_widget/change_water_modal/content_modal/content_change_time_modal.dart';
