+++ lib/widgets/water_widget/change_water_modal/title_change_time_modal.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
-import 'package:iot_plant_control/controller/water_controller.dart';
+import 'package:iot_plant_control/controller/watering_controller/water_controller.dart';
 import 'package:iot_plant_control/models/water_time.dart';
 
 class TitleChangeTimeModal extends StatelessWidget {
